chore: move blueprintConfig migrations
