+++ meteor/server/migration/lib.ts
@@ -4,9 +4,8 @@ import {
 	MigrationStepInput,
 	MigrationStepInputFilteredResult,
 	MigrationStepBase,
-	IBlueprintConfig,
 } from 'tv-automation-sofie-blueprints-integration'
-import { Collections, objectPathGet, DBObj, ProtectedString, objectPathSet } from '../../lib/lib'
+import { Collections, objectPathGet, DBObj, ProtectedString } from '../../lib/lib'
 import { Meteor } from 'meteor/meteor'
 import { PeripheralDevices } from '../../lib/collections/PeripheralDevices'
 import { PeripheralDeviceAPI } from '../../lib/api/peripheralDevice'
@@ -249,36 +248,6 @@ export function renamePropertiesInCollection<T extends DBInterface, DBInterface 
 				collection.update(doc._id, doc)
 			})
 			//
-		},
-	}
-}
-
-export function migrateConfigToBlueprintConfig<
-	T extends DBInterface,
-	DBInterface extends { _id: ProtectedString<any>; blueprintConfig: IBlueprintConfig }
->(id: string, collection: TransformedCollection<T, DBInterface>) {
-	return {
-		id,
-		canBeRunAutomatically: true,
-		validate: () => {
-			const documents = collection.find({ config: { $exists: true } }).fetch()
-			if (documents.length) {
-				return true
-			}
-			return false
-		},
-		migrate: () => {
-			const documents = collection.find({ config: { $exists: true } }).fetch() as Array<
-				T & { config: Array<{ _id: string; value: any }> }
-			>
-			for (const document of documents) {
-				document.blueprintConfig = {}
-				for (const item of document.config) {
-					objectPathSet(document.blueprintConfig, item._id, item.value)
-				}
-				delete document.config
-				collection.update(document._id, document)
-			}
 		},
 	}
 }
